Quick Look at a Company's Financials

This function allows a user to view several financial ratios for a company.

filter_company_data(tik)

Arguments

tik

A character string consisting of the stock ticker for the company.

Value

A dataframe containing financial ratios for the company if it is a public stock on NASDAQ or NYSE.

Examples

filter_company_data(tik = "AAPL")
#> --------------------------------------
#> -- Starting data retrieval from EDGAR
#> -- May take ~30sec
#> --------------------------------------
#> -- Getting Revenue
#> -- Getting Operating Income
#> -- Getting Net Income
#> -- Getting Gross Profit
#> -- Tidying up the data
#> taxonomyccpuomdata.cikdata.entityNamedata.locdata.startdata.endyearOperatingIncomeLossGrossProfitrevenuenet_income
#> Got the financials of 10882 companies